Fix: channel owner changes when someone else left

+++ events/voiceStateUpdate.js
@@ -34,8 +34,8 @@ module.exports = {
 
 		if (
 			newState.channel &&
-      creatorChannels.includes(newState.channel.id) &&
-      newState.channel.members.size === 1
+			creatorChannels.includes(newState.channel.id) &&
+			newState.channel.members.size === 1
 		) {
 			const guild = newState.guild;
 			const user = newState.member.user;
@@ -90,56 +90,61 @@ module.exports = {
 
 		if (
 			oldState.channel &&
-      oldState.member.user.id === newState.member.user.id
+			oldState.member.user.id === newState.member.user.id
 		) {
 			if (
 				oldState.channel.members.size === 0 &&
-        userChannels[oldState.channel.id]
+				userChannels[oldState.channel.id]
 			) {
-				const creatorChannel = newState.guild.channels.cache.get(
-					userChannels[oldState.channel.id].creatorChannelId,
-				);
-				if (creatorChannel) {
-					try {
-						delete userChannels[oldState.channel.id];
-						await oldState.channel.delete();
-					}
-					catch (error) {
-						console.error(`Failed to delete channel: ${error.message}`);
+				const channelData = userChannels[oldState.channel.id];
+				if (channelData.channelOwnerId === oldState.member.user.id) {
+					const creatorChannel = newState.guild.channels.cache.get(
+						channelData.creatorChannelId,
+					);
+					if (creatorChannel) {
+						try {
+							delete userChannels[oldState.channel.id];
+							await oldState.channel.delete();
+						}
+						catch (error) {
+							console.error(`Failed to delete channel: ${error.message}`);
+						}
 					}
 				}
 			}
 			else if (
 				oldState.channel.members.size > 0 &&
-        userChannels[oldState.channel.id] &&
-        !oldState.selfMute &&
-        !oldState.selfDeaf &&
-		!oldState.selfStream &&
-        !newState.selfMute &&
-        !newState.selfDeaf &&
-		!newState.selfStream
+				userChannels[oldState.channel.id] &&
+				!oldState.selfMute &&
+				!oldState.selfDeaf &&
+				!oldState.selfStream &&
+				!newState.selfMute &&
+				!newState.selfDeaf &&
+				!newState.selfStream
 			) {
 				const nextMember = oldState.channel.members.first();
 				const channelData = userChannels[oldState.channel.id];
 
-				await oldState.channel.permissionOverwrites.edit(nextMember, {
-					ManageChannels: true,
-					ManageRoles: true,
-				});
+				if (channelData.channelOwnerId === oldState.member.user.id) {
+					await oldState.channel.permissionOverwrites.edit(nextMember, {
+						ManageChannels: true,
+						ManageRoles: true,
+					});
 
-				await oldState.channel.permissionOverwrites.delete(oldState.member);
+					await oldState.channel.permissionOverwrites.delete(oldState.member);
 
-				let newChannelName = `${nextMember.displayName}'s VC`;
-				if (channelData.isLocked) {
-					newChannelName = `Locked | ${nextMember.displayName}'s VC`;
+					let newChannelName = `${nextMember.displayName}'s VC`;
+					if (channelData.isLocked) {
+						newChannelName = `Locked | ${nextMember.displayName}'s VC`;
+					}
+
+					await oldState.channel.edit({ name: newChannelName });
+
+					data.userChannels[oldState.channel.id].channelOwnerId = nextMember.id;
+					data.userChannels[oldState.channel.id].channelOwnerName =
+						nextMember.user.username;
+					data.userChannels[oldState.channel.id].channelName = newChannelName;
 				}
-
-				await oldState.channel.edit({ name: newChannelName });
-
-				data.userChannels[oldState.channel.id].channelOwnerId = nextMember.id;
-				data.userChannels[oldState.channel.id].channelOwnerName =
-          nextMember.user.username;
-				data.userChannels[oldState.channel.id].channelName = newChannelName;
 			}
 		}
